requirement.txt
pymongo
SpeechRecognition
google-generativeai
pyttsx3
pypiwin32
streamlit
